requirements.txt
pymongo==3.11.4
Flask==1.1.2
python_bcrypt==0.3.2
gunicorn